move createFaker to base.
