+++ generators/cypress/generator.mts
@@ -18,7 +18,7 @@
  */
 import _ from 'lodash';
 
-import { stringHashCode } from '../base/support/index.mjs';
+import { stringHashCode, createFaker } from '../base/support/index.mjs';
 import BaseApplicationGenerator, { type Entity } from '../base-application/index.mjs';
 import { cypressFiles, cypressEntityFiles } from './files.mjs';
 import { clientFrameworkTypes } from '../../jdl/jhipster/index.mjs';
@@ -28,7 +28,6 @@ import { GENERATOR_CYPRESS, GENERATOR_BOOTSTRAP_APPLICATION } from '../generator
 import type { CypressApplication } from './types.mjs';
 import { generateTestEntity as entityWithFakeValues } from '../client/support/index.mjs';
 import { BaseApplicationGeneratorDefinition } from '../base-application/tasks.mjs';
-import { createFaker } from '../bootstrap-application-base/faker.mjs';
 
 const { ANGULAR } = clientFrameworkTypes;
