+++ generators/docker-compose/generator.mjs
@@ -36,8 +36,7 @@ import {
   searchEngineTypes,
 } from '../../jdl/jhipster/index.mjs';
 import { GENERATOR_DOCKER_COMPOSE } from '../generator-list.mjs';
-import { stringHashCode } from '../base/support/index.mjs';
-import { createFaker } from '../bootstrap-application-base/faker.mjs';
+import { stringHashCode, createFaker } from '../base/support/index.mjs';
 
 const { GATEWAY, MONOLITH } = applicationTypes;
 const { PROMETHEUS } = monitoringTypes;
